requirements.txt
google-play-scraper
pandas
google-generativeai
python-dotenv
streamlit
plotly
